+++ KDT0_JeongBeomHwan/public/index.html
@@ -26,13 +26,13 @@
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@48,400,0,0" />
 
     <!-- JS -->
-    <script defer src="./javascripts/index.js"></script>
-    <script defer type="module" src="./javascripts/firebase.js"></script>
+
+    <script defer type="module" src="./javascripts/login.js"></script>
   </head>
   <body>
     <section id="login-form">
       <img src="./images/login-form_main-logo.png" alt="Main-Logo" />
-      <a><button>구글 계정으로 시작</button></a>
+      <a><button class="login-btn">구글 계정으로 시작</button></a>
     </section>
   </body>
 </html>
