+++ KDT0_JeongBeomHwan/public/main.html
@@ -34,7 +34,7 @@
     ></script>
 
     <!-- JS -->
-    <script defer src="./javascripts/main.js"></script>
+    <script defer type="module" src="./javascripts/main.js"></script>
   </head>
   <body class="mo">
     <div id="DIM" class="none"></div>
@@ -118,7 +118,7 @@
       <img class="header-logo" src="./images/main_logo_width-long.png" alt="logo" />
       <div class="header-userinfo">
         <span class="header-userinfo_greeting">안녕하세요 님</span>
-        <a class="header-userinfo_logout" href="./index.html">로그아웃</a>
+        <a class="header-userinfo_logout">로그아웃</a>
       </div>
     </header>
     <!--
